ADKStarterKit/ui/requirements.txt
google-adk>=0.3.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0   